feat: adicionar lógica de validação de usuário na exclusão de reservas

+++ src/modules/delete_booking/app/delete_booking_controller.py
@@ -19,9 +19,15 @@ class DeleteBookingController:
         # TODO: Fazer a lógica de receber o token e buscar o usuário. Mandar as infos do usuário para o usecase 
 
         try:
+
+            if request.data.get('user_from_authorizer') is None:
+                raise MissingParameters('user authorizer')
+            
+            user = request.data.get('user_from_authorizer')
+
             if request.data.get('booking_id') is None:
                 raise MissingParameters('booking_id')
-            booking = self.usecase(booking_id=request.data.get('booking_id'))
+            booking = self.usecase(booking_id=request.data.get('booking_id'),  user=user)
             viewmodel = DeleteBookingViewModel(booking)
             
             return OK(viewmodel.to_dict())
